complete up to exercise 4-22

+++ controllers/blogs.js
@@ -1,14 +1,26 @@
 const blogsRouter = require('express').Router()
-// const { response } = require('../app')
 const Blog = require('../models/blog')
+const jwt = require('jsonwebtoken')
+const User = require('../models/user')
+// const user = require('../models/user')
+
+// const getTokenFromRequest = (request) => {
+// 	const authorization = request.get('authorization')
+// 	if (authorization && authorization.startWith('Bearer ')) {
+// 		return authorization.replace('Bearer ', '')
+// 	}
+// 	return null
+// }
 
 blogsRouter.get('/', async (request, response) => {
-	const blogs = await Blog.find({})
+	const blogs = await Blog
+		.find({}).populate('user', {username: 1, name:1})
 	response.json(blogs)
 })
 
 blogsRouter.get('/:id', async (request, response) => {
-	const blog = await Blog.findById(request.params.id)
+	const blog = await Blog
+		.findById(request.params.id).populate('user', {username:1, name:1})
 	if (blog) {
 		response.json(blog)
 	} else {
@@ -18,19 +30,32 @@ blogsRouter.get('/:id', async (request, response) => {
   
 blogsRouter.post('/', async (request, response) => {
 	const body = request.body
-	// console.log('post blog', body)
 	if (!body.title || !body.url) {
-		return response.status(400).end()
+		return response.status(400).json({ error: 'title and url are required'})
 	}
+
+	const decodedToken = jwt.verify(request.token, process.env.SECRET)
+	if (!decodedToken.id) {
+		return response.status(401).json({ error: 'token invalid'})
+	}
+	const user = await User.findById(request.user)
+	console.log('post blog user', request.user)
+	// console.log('post blog user', user._id)
 
 	const blog = new Blog({
 		title: body.title,
 		author: body.author,
 		url: body.url,
-		likes: !body.likes ? 0 : body.likes
+		likes: !body.likes ? 0 : body.likes,
+		user: user._id,
 	})
 
+	console.log('post blog', blog)
+
 	const savedBlog = await blog.save()
+	user.blogs = user.blogs.concat(savedBlog._id)
+	await user.save()
+
 	response.status(201).json(savedBlog)
 })
 
@@ -51,6 +76,21 @@ blogsRouter.put('/:id', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+
+	const decodedToken = jwt.verify(request.token, process.env.SECRET)
+	if (!decodedToken.id) {
+		return response.status(401).json({ error: 'token invalid'})
+	}
+	const user = request.user
+	const blog = await Blog.findById(request.params.id)
+
+	console.log('delete blog user', user.toString())
+	console.log('delete blog blog', blog.user.toString())
+
+	if (!(blog.user.toString() === user.toString())) {
+		return response.status(401).json({ error: 'unable to delete this blog post'})
+	}
+
 	await Blog.findByIdAndRemove(request.params.id)
 	response.status(204).end()
 })
